Update README.md
A basic website showing the information related to covid-19 such as symptoms and precautions that awares the people regarding the safety measures they can take to be safe from the virus. A dashboard also linked with the website wich shows data related to covid-19 like numbers of cases , recoveries and death due to covid-19 virus  across the globe.